🚧 update dashboard ui #29
